Refactor away from BuildState class

+++ packages/cli/src/commands/build.ts
@@ -1,16 +1,25 @@
-import { Env, LoadedEnvFiles, loadEnvConfig } from '@next/env';
+import { loadEnvConfig } from '@next/env';
+import { isAbsolute } from '@sentry/utils';
 import {
   execCommand,
   getScriptName,
   scanParentDirs,
   spawnAsync,
 } from '@vercel/build-utils';
+import chalk from 'chalk';
+import { SpawnOptions } from 'child_process';
+import { assert } from 'console';
 import fs from 'fs-extra';
 import { join } from 'path';
+import pluralize from 'pluralize';
 import Client from '../util/client';
 import getArgs from '../util/get-args';
 import handleError from '../util/handle-error';
 import { isSettingValue } from '../util/is-setting-value';
+import cmd from '../util/output/cmd';
+import code from '../util/output/code';
+import param from '../util/output/param';
+import stamp from '../util/output/stamp';
 import { getCommandName } from '../util/pkg-name';
 import { findFramework } from '../util/projects/find-framework';
 import { VERCEL_DIR } from '../util/projects/link';
@@ -20,15 +29,6 @@ import {
 } from '../util/projects/project-settings';
 import pull from './pull';
 
-import { assert } from 'console';
-import { isAbsolute } from '@sentry/utils';
-import { SpawnOptions } from 'child_process';
-import cmd from '../util/output/cmd';
-import param from '../util/output/param';
-import code from '../util/output/code';
-import { Framework } from '../../../frameworks/dist/types';
-import pluralize from 'pluralize';
-
 const help = () => {
   // @todo help output
   return console.log('vercel build');
@@ -36,6 +36,17 @@ const help = () => {
 
 const OUTPUT_DIR = '.output';
 const VERCEL_PLUGIN_PREFIX = 'vercel-plugin-';
+
+const fields: {
+  name: string;
+  value: keyof ProjectLinkAndSettings['settings'];
+}[] = [
+  { name: 'Directory Listing', value: 'directoryListing' },
+  { name: 'Build Command', value: 'buildCommand' },
+  { name: 'Output Directory', value: 'outputDirectory' },
+  { name: 'Root Directory', value: 'rootDirectory' },
+  { name: 'Development Command', value: 'devCommand' },
+];
 
 export default async function main(client: Client) {
   let argv;
@@ -62,59 +73,119 @@ export default async function main(client: Client) {
     }
     project = await readProjectSettings(join(cwd, VERCEL_DIR));
   }
-  client.output.print(getCommandName('build') + '\n');
-  const buildState = new BuildState(cwd, client, project!);
+
+  cwd = project.settings.rootDirectory
+    ? join(cwd, project.settings.rootDirectory)
+    : cwd;
 
   // Load the environment
-  await buildState.loadEnv(false);
-
-  buildState.renderConfig();
+  const { combinedEnv, loadedEnvFiles } = loadEnvConfig(cwd, false, {
+    info: () => ({}), // we don't want to log this yet.
+    error: (...args: any[]) => client.output.error(args.join(' ')),
+  });
 
   const spawnOpts = {
-    env: buildState.combinedEnv,
+    env: { ...combinedEnv, VERCEL: '1' },
   };
 
-  process.chdir(buildState.cwd);
+  process.chdir(cwd);
 
+  const framework = findFramework(project.settings.framework);
+  const buildState = { ...project.settings };
+  if (framework) {
+    client.output.log(`Retrieved Project Settings:`);
+    client.output.print(
+      chalk.dim(`  - ${chalk.bold(`Framework Preset:`)} ${framework.name}\n`)
+    );
+
+    for (let field of fields) {
+      const defaults = (framework.settings as any)[field.value];
+      if (defaults) {
+        client.output.print(
+          chalk.dim(
+            `  - ${chalk.bold(`${field.name}:`)} ${`${
+              project.settings[field.value]
+                ? project.settings[field.value] + ` (override)`
+                : isSettingValue(defaults)
+                ? defaults.value
+                : chalk.italic(`${defaults.placeholder}`)
+            }`}\n`
+          )
+        );
+      }
+      (buildState as any)[field.value] = project.settings[field.value]
+        ? project.settings[field.value]
+        : defaults
+        ? isSettingValue(defaults)
+          ? defaults.value
+          : null
+        : null;
+    }
+  }
+
+  if (loadedEnvFiles.length > 0) {
+    client.output.log(
+      `Loaded Environment Variables from ${loadedEnvFiles.length} ${pluralize(
+        'file',
+        loadedEnvFiles.length
+      )}:`
+    );
+    for (let envFile of loadedEnvFiles) {
+      client.output.print(chalk.dim(`  - ${envFile.path}\n`));
+    }
+  }
+
+  // Load plugins
   let plugins;
   try {
-    plugins = await loadCliPlugins(client, buildState.cwd);
+    plugins = await loadCliPlugins(client, cwd);
   } catch (error) {
-    client.output.error('Failed to load plugins');
+    client.output.error('Failed to load CLI Plugins');
     handleError(error);
     return 1;
   }
 
-  if (plugins) {
+  if (plugins?.pluginCount && plugins?.pluginCount > 0) {
     client.output.log(
-      `Found ${plugins.length} ${pluralize('plugin', plugins.length)}`
+      `Loaded ${plugins.pluginCount} CLI ${pluralize(
+        'Plugin',
+        plugins.pluginCount
+      )}`
     );
-    for (let item of plugins) {
-      const { name, plugin } = item;
-      if (typeof plugin.preBuild === 'function') {
-        client.output.log(`Running ${code(name + '.preBuild')}`);
-        try {
-          await plugin.preBuild();
-        } catch (error) {
-          client.output.error(`${code(name + '.preBuild')} failed`);
-          handleError(error);
-          return 1;
+    // preBuild Plugins
+    if (plugins.preBuildPlugins.length > 0) {
+      client.output.log(
+        `Running ${plugins.pluginCount} CLI ${pluralize(
+          'Plugin',
+          plugins.pluginCount
+        )} before Build Command:`
+      );
+      for (let item of plugins.preBuildPlugins) {
+        const { name, plugin } = item;
+        if (typeof plugin.preBuild === 'function') {
+          const pluginStamp = stamp();
+          client.output.log(`Running ${code(name + '.preBuild')}`);
+          try {
+            await plugin.preBuild();
+            client.output.log(
+              `Completed ${code(name + '.preBuild')} ${chalk.dim(
+                `${pluginStamp()}`
+              )}`
+            );
+          } catch (error) {
+            client.output.error(`${code(name + '.preBuild')} failed`);
+            handleError(error);
+            return 1;
+          }
         }
       }
     }
   }
 
   let result: boolean;
-  if (typeof buildState.buildCommand?.value === 'string') {
-    buildState.client.output.log(
-      `Running ${cmd(buildState.buildCommand.value)} (${
-        buildState.buildCommand.source == BuildStateSource.DASHBOARD
-          ? 'project settings'
-          : 'package.json'
-      })\n`
-    );
-
-    result = await execCommand(buildState.buildCommand.value, {
+  if (typeof buildState.buildCommand === 'string') {
+    client.output.log(`Running Build Command: ${cmd(buildState.buildCommand)}`);
+    result = await execCommand(buildState.buildCommand, {
       ...spawnOpts,
       // Yarn v2 PnP mode may be activated, so force
       // "node-modules" linker style
@@ -136,18 +207,18 @@ export default async function main(client: Client) {
   if (!result) {
     client.output.error(
       `Missing required "${cmd(
-        buildState.buildCommand?.value || 'vercel-build' || 'build'
-      )}" script in ${param(buildState.cwd)}"\n`
+        buildState.buildCommand || 'vercel-build' || 'build'
+      )}" script in ${param(cwd)}"\n`
     );
     return 1;
   }
 
   client.output.log(
     `Moving output from ${param(
-      buildState.outputDirectory?.value || '.next'
+      buildState.outputDirectory || '.next'
     )} to ${param(OUTPUT_DIR)}`
   );
-  if (typeof buildState.outputDirectory?.value !== 'string') {
+  if (typeof buildState.outputDirectory !== 'string') {
     client.output.error(
       `Could not determine output directory. Please run ${getCommandName(
         'pull'
@@ -156,27 +227,44 @@ export default async function main(client: Client) {
     return 1;
   }
 
-  // Move to .output
-  await fs.remove(join(cwd, OUTPUT_DIR));
-  await fs.move(
-    join(cwd, buildState.outputDirectory.value),
-    join(cwd, OUTPUT_DIR)
-  );
+  const maybeOutputDir = await framework?.getOutputDirName(cwd);
+  const realOutputDir = buildState.outputDirectory
+    ? join(cwd, buildState.outputDirectory)
+    : maybeOutputDir;
+
+  if (realOutputDir) {
+    await fs.pathExists(realOutputDir);
+    // Move to .output
+    await fs.remove(join(cwd, OUTPUT_DIR));
+    await fs.move(realOutputDir, join(cwd, OUTPUT_DIR));
+  }
 
   client.output.log(
-    `Moved ${param(buildState.outputDirectory.value || '.next')} to ${param(
+    `Moved ${param(buildState.outputDirectory || '.next')} to ${param(
       OUTPUT_DIR
     )}`
   );
 
-  // Plugins
-  if (plugins) {
-    for (let item of plugins) {
+  // Build Plugins
+  if (plugins?.buildPlugins && plugins.buildPlugins.length > 0) {
+    client.output.log(
+      `Running ${plugins.pluginCount} CLI ${pluralize(
+        'Plugin',
+        plugins.pluginCount
+      )} after Build Command`
+    );
+    for (let item of plugins.buildPlugins) {
       const { name, plugin } = item;
       if (typeof plugin.build === 'function') {
+        const pluginStamp = stamp();
         client.output.log(`Running ${code(name + '.build')}`);
         try {
           await plugin.build();
+          client.output.log(
+            `Completed ${code(name + '.build')} ${chalk.dim(
+              `${pluginStamp()}`
+            )}`
+          );
         } catch (error) {
           client.output.error(`${code(name + '.build')} failed`);
           handleError(error);
@@ -189,161 +277,6 @@ export default async function main(client: Client) {
   client.output.success('Build completed');
 
   return 0;
-}
-
-export enum BuildStateSource {
-  DEFAULT,
-  DASHBOARD,
-}
-
-export interface BuildStateRecord {
-  value: string | null;
-  source: BuildStateSource;
-}
-
-export class BuildState {
-  // Current working directory
-  cwd: string;
-  // Vercel client
-  client: Client;
-
-  framework?: Framework;
-
-  // Output directory
-  outputDirectory?: BuildStateRecord;
-  // Build command
-  buildCommand?: BuildStateRecord;
-  // Root directory
-  rootDirectory?: BuildStateRecord;
-  // Combined environment variables
-  combinedEnv?: Env;
-  // Loaded .env files
-  loadedEnvFiles?: LoadedEnvFiles;
-  // Project Link + settings
-  project: ProjectLinkAndSettings;
-
-  constructor(cwd: string, client: Client, project: ProjectLinkAndSettings) {
-    this.cwd = cwd;
-    this.client = client;
-    this.project = project;
-
-    if (typeof project.settings.rootDirectory === 'string') {
-      this.setCwd(join(cwd, project.settings.rootDirectory));
-      this.setRootDirectory(
-        project.settings.rootDirectory,
-        BuildStateSource.DASHBOARD
-      );
-    } else {
-      this.setRootDirectory(null, BuildStateSource.DEFAULT);
-    }
-
-    this.framework = findFramework(project.settings.framework);
-
-    let buildCommand = project.settings.buildCommand;
-
-    if (typeof buildCommand === 'string') {
-      this.setBuildCommand(buildCommand, BuildStateSource.DASHBOARD);
-    } else if (this.framework) {
-      const defaults = this.framework.settings.buildCommand;
-      if (isSettingValue(defaults)) {
-        buildCommand = defaults.value;
-        this.setBuildCommand(buildCommand, BuildStateSource.DEFAULT);
-      } else {
-        this.setBuildCommand(null, BuildStateSource.DEFAULT);
-      }
-    }
-
-    let outputDirectory = project.settings.outputDirectory;
-    if (typeof outputDirectory === 'string') {
-      this.setOutputDirectory(outputDirectory, BuildStateSource.DASHBOARD);
-    } else if (this.framework) {
-      const defaults = this.framework.settings.outputDirectory;
-      if (isSettingValue(defaults)) {
-        outputDirectory = defaults.value;
-        this.setOutputDirectory(outputDirectory, BuildStateSource.DEFAULT);
-      } else {
-        // the default output directory is actually .next
-        // @todo need to handle Other which is `public` if it exists or `.`
-        this.setOutputDirectory('.next', BuildStateSource.DEFAULT);
-      }
-    }
-  }
-
-  setCwd(cwd: string) {
-    this.cwd = cwd;
-  }
-
-  setBuildCommand(value: BuildStateRecord['value'], source: BuildStateSource) {
-    this.buildCommand = {
-      source,
-      value,
-    };
-  }
-
-  setOutputDirectory(
-    value: BuildStateRecord['value'],
-    source: BuildStateSource
-  ) {
-    this.outputDirectory = {
-      source,
-      value,
-    };
-  }
-
-  setRootDirectory(value: BuildStateRecord['value'], source: BuildStateSource) {
-    this.rootDirectory = {
-      source,
-      value,
-    };
-  }
-
-  async loadEnv(dev: boolean = false) {
-    const { combinedEnv, loadedEnvFiles } = await loadEnvConfig(this.cwd, dev);
-    this.combinedEnv = { ...combinedEnv, VERCEL: '1' };
-    this.loadedEnvFiles = loadedEnvFiles;
-  }
-
-  // print config pretty prints the build state, it's goal is to remove
-  // imperative logging/printing throughout the `main` function
-  renderConfig() {
-    if (this.framework) {
-      this.client.output.log(
-        `Framework preset: ${param(this.framework.name)} (project settings)`
-      );
-    }
-
-    if (typeof this.outputDirectory?.value === 'string') {
-      this.client.output.log(
-        `Output directory: ${param(this.outputDirectory.value)} (${
-          this.outputDirectory.source == BuildStateSource.DASHBOARD
-            ? 'project settings'
-            : 'framework default'
-        })`
-      );
-    }
-
-    if (typeof this.rootDirectory?.value === 'string') {
-      this.client.output.log(
-        `Root directory: ${param(this.rootDirectory.value)} (${
-          this.rootDirectory.source == BuildStateSource.DASHBOARD
-            ? 'project settings'
-            : 'default'
-        })`
-      );
-    }
-
-    this.client.output.log(
-      `Build command: ${param(
-        this.buildCommand?.value ||
-          this.framework?.settings.buildCommand.placeholder ||
-          ''
-      )} (${
-        this.buildCommand?.source == BuildStateSource.DASHBOARD
-          ? 'project settings'
-          : 'default'
-      })`
-    );
-  }
 }
 
 export async function runPackageJsonScript(
@@ -386,7 +319,7 @@ export async function runPackageJsonScript(
     }
   }
 
-  client.output.log(`Running ${cmd(opts.prettyCommand)} (package.json)\n`);
+  client.output.log(`Running Build Command: ${cmd(opts.prettyCommand)}\n`);
   await spawnAsync(cliType, ['run', scriptName], opts);
   client.output.print('\n'); // give it some room
   client.output.debug(`Script complete [${Date.now() - runScriptTime}ms]`);
@@ -396,29 +329,41 @@ export async function runPackageJsonScript(
 async function loadCliPlugins(client: Client, cwd: string) {
   const { packageJson } = await scanParentDirs(cwd, true);
   if (packageJson) {
-    const plugins = [];
+    let pluginCount = 0;
+    const preBuildPlugins = [];
+    const buildPlugins = [];
+
     const deps = [
       ...Object.keys(packageJson.dependencies || {}),
       ...Object.keys(packageJson.devDependencies || {}),
     ].filter(dep => dep.startsWith(VERCEL_PLUGIN_PREFIX));
 
     for (let dep of deps) {
+      pluginCount++;
       const resolved = require.resolve(dep, {
         paths: [cwd, process.cwd()],
       });
       let plugin;
       try {
         plugin = require(resolved);
-        plugins.push({
-          plugin,
-          name: dep,
-        });
+        if (typeof plugin.preBuild === 'function') {
+          preBuildPlugins.push({
+            plugin,
+            name: dep,
+          });
+        }
+        if (typeof plugin.build === 'function') {
+          buildPlugins.push({
+            plugin,
+            name: dep,
+          });
+        }
       } catch (error) {
         client.output.error(`Failed to import ${code(dep)}`);
         throw error;
       }
     }
 
-    return plugins;
+    return { pluginCount, preBuildPlugins, buildPlugins };
   }
 }
